% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum.R
\name{somma}
\alias{somma}
\title{somma}
\usage{
somma(a, b)
}
\arguments{
\item{a}{\link{numeric} first addendum}

\item{b}{\link{numeric} second}
}
\value{
\link{numeric} the sum of the two inputs
}
\description{
This function sums two real numbers.
}
\examples{
somma(3, 2)
}
